Cleam sampling tests
